// src/API.ts
export type Post = {
  __typename: "Post",
  id: string,
  title: string,
  content: string,
  image?: string | null,
  votes?: ModelVoteConnection | null,
  comments?: ModelCommentConnection | null,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};

export type ModelVoteConnection = {
  __typename: "ModelVoteConnection",
  items:  Array<Vote | null >,
  nextToken?: string | null,
};

export type Vote = {
  __typename: "Vote",
  id: string,
  vote: string,
  postID: string,
  post?: Post | null,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};

export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items:  Array<Comment | null >,
  nextToken?: string | null,
};

export type Comment = {
  __typename: "Comment",
  id: string,
  postID: string,
  post?: Post | null,
  content: string,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};
// ...
export type CreateVoteInput = {
  id?: string | null,
  vote: string,
  postID: string,
};

export type ModelVoteConditionInput = {
  vote?: ModelStringInput | null,
  postID?: ModelIDInput | null,
  and?: Array< ModelVoteConditionInput | null > | null,
  or?: Array< ModelVoteConditionInput | null > | null,
  not?: ModelVoteConditionInput | null,
};

export type UpdateVoteInput = {
  id: string,
  vote?: string | null,
  postID?: string | null,
};
// ...
export type CreateVoteMutation = {
  createVote?:  {
    __typename: "Vote",
    id: string,
    vote: string,
    postID: string,
    post?:  {
      __typename: "Post",
      id: string,
      title: string,
      content: string,
      image?: string | null,
      votes?:  {
        __typename: "ModelVoteConnection",
        items:  Array< {
          __typename: "Vote",
          id: string,
          vote: string,
          postID: string,
          createdAt: string,
          updatedAt: string,
          owner?: string | null,
        } | null >,
        nextToken?: string | null,
      } | null,
      comments?:  {
        __typename: "ModelCommentConnection",
        items:  Array< {
          __typename: "Comment",
          id: string,
          postID: string,
          content: string,
          createdAt: string,
          updatedAt: string,
          owner?: string | null,
        } | null >,
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      owner?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    owner?: string | null,
  } | null,
};

export type UpdateVoteMutationVariables = {
  input: UpdateVoteInput,
  condition?: ModelVoteConditionInput | null,
};

export type UpdateVoteMutation = {
  updateVote?:  {
    __typename: "Vote",
    id: string,
    vote: string,
    postID: string,
    post?:  {
      __typename: "Post",
      id: string,
      title: string,
      content: string,
      image?: string | null,
      votes?:  {
        __typename: "ModelVoteConnection",
        items:  Array< {
          __typename: "Vote",
          id: string,
          vote: string,
          postID: string,
          createdAt: string,
          updatedAt: string,
          owner?: string | null,
        } | null >,
        nextToken?: string | null,
      } | null,
      comments?:  {
        __typename: "ModelCommentConnection",
        items:  Array< {
          __typename: "Comment",
          id: string,
          postID: string,
          content: string,
          createdAt: string,
          updatedAt: string,
          owner?: string | null,
        } | null >,
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      owner?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    owner?: string | null,
  } | null,
};